refactor api key store file helpers
